recipe-crawler/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
pandas==2.0.3
supabase==2.0.2
python-dotenv==1.0.0 